Update patch set 1
Patch Set 1:

Is there a reason we set for passion and bravo targets the handset and headset min audio gain so high?
It makes it difficult to have a private conversation (handset value) and it blows your ears off (headset values).
Should we consider putting these both handset and headset values to -1900/100?

Patch-set: 1